Add support for 2's complement integer encoding with 1-8 bytes width. Add unit tests for integer encoding. Add unit tests for string encoding. Change atom processor interface in order to have stateless processors.

+++ core/src/main/java/org/beyene/protege/data/Primitive.java
@@ -26,7 +26,7 @@ public final class Primitive<T> {
 
 	public static final Primitive<Boolean> BOOLEAN = new Primitive<>(Boolean.class, Type.BOOLEAN);
 	public static final Primitive<Byte[]> BYTES = new Primitive<>(Byte[].class, Type.BYTE);
-	public static final Primitive<Integer> INTEGER = new Primitive<>(Integer.class, Type.INTEGER);
+	public static final Primitive<Long> INTEGER = new Primitive<>(Long.class, Type.INTEGER);
 	public static final Primitive<String> STRING = new Primitive<>(String.class, Type.STRING);
 	public static final Primitive<Float> FLOAT = new Primitive<>(Float.class, Type.FLOAT);
 	public static final Primitive<Double> DOUBLE = new Primitive<>(Double.class, Type.DOUBLE);
